create runner package to separate concerns
The idea here is that bundle.go should only deal with CLI aspects:
parameter parsing, printing help, etc.
